requirements.txt
# Crypto Trading Bot Dependencies
# Install with: pip install -r requirements.txt

# Exchange connectivity
ccxt>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0

# Optional: for visualization (backtesting)
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: for performance optimization
numba>=0.57.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
